refactor(tests): Rewrite part of the test in TS
