+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@mogeko/gatsby-remark-plantuml-lite",
+  "name": "gatsby-remark-plantuml-lite",
   "version": "0.0.3-alpha.3",
   "description": "A Gatsby plugin to transform PlantUML code blocks into SVG images.",
   "main": "dist/index.js",
@@ -26,11 +26,12 @@
   "scripts": {
     "build": "tsc",
     "release": "npm run build && npm publish",
-    "test": "npm run build && jest ./tests --watch",
-    "coverage": "npm run build && jest --coverage"
+    "test": "jest ./tests --watch",
+    "coverage": "jest --coverage"
   },
   "devDependencies": {
     "@babel/runtime": "^7.12.5",
+    "@types/jest": "^26.0.15",
     "@types/node": "^14.14.7",
     "@types/yargs-parser": "^15.0.0",
     "@typescript-eslint/eslint-plugin": "^4.7.0",
@@ -39,6 +40,7 @@
     "gatsby-plugin-utils": "^0.3.0",
     "jest": "^26.6.3",
     "remark": "^13.0.0",
+    "ts-jest": "^26.4.4",
     "typescript": "^4.0.5"
   },
   "dependencies": {
