+++ tests/lib.test.ts
@@ -1,6 +1,5 @@
-const fs = require("fs")
-const remark = require("remark")
-const nodeOperator = require("../dist/lib")
+import remark = require("remark")
+import nodeOperator from "../src/lib"
 
 test("基础测试", () => {
   const raw = remark().parse(`
@@ -10,10 +9,10 @@ test("基础测试", () => {
   @enduml
   \`\`\`
   `)
-  const example = JSON.parse(fs.readFileSync("./tests/examples/test.data.0.json"))
-  const result = JSON.parse(JSON.stringify(nodeOperator.default(raw, (encoded) => {
+  const example = require("./examples/test.data.0.json")
+  const result = nodeOperator(raw, (encoded) => {
     return `https://www.plantuml.com/plantuml/svg/${encoded}`
-  })))
+  })
   expect(result).toStrictEqual(example)
 })
 
@@ -25,11 +24,10 @@ test("测试编码是否正确", () => {
   @enduml
   \`\`\`
   `)
-  const url = nodeOperator.default(raw, (encoded) => {
+  nodeOperator(raw, (encoded) => {
     expect(encoded).toStrictEqual("SoWkIImgAStDuN9KqBLJSB9Iy4ZDoSbNq5TuidV1qwLxrRaSKlDIWF80")
     return encoded
-  }).children[0].children[0].url
-  expect(url).toStrictEqual("SoWkIImgAStDuN9KqBLJSB9Iy4ZDoSbNq5TuidV1qwLxrRaSKlDIWF80")
+  })
 })
 
 test("更复杂的案例", () => {
@@ -46,10 +44,10 @@ test("更复杂的案例", () => {
   console.log("TEST")
   \`\`\`
   `)
-  const example = JSON.parse(fs.readFileSync("./tests/examples/test.data.1.json"))
-  const result = JSON.parse(JSON.stringify(nodeOperator.default(raw, (encoded) => {
+  const example = require("./examples/test.data.1.json")
+  const result = nodeOperator(raw, (encoded) => {
     return `https://www.plantuml.com/plantuml/svg/${encoded}`
-  })))
+  })
   expect(result).toStrictEqual(example)
 })
 
@@ -61,9 +59,9 @@ test("Markdown 的嵌套", () => {
   > @enduml
   > \`\`\`
   `)
-  const example = JSON.parse(fs.readFileSync("./tests/examples/test.data.2.json"))
-  const result = JSON.parse(JSON.stringify(nodeOperator.default(raw, (encoded) => {
+  const example = require("./examples/test.data.2.json")
+  const result = nodeOperator(raw, (encoded) => {
     return `https://www.plantuml.com/plantuml/svg/${encoded}`
-  })))
+  })
   expect(result).toStrictEqual(example)
 })
