+++ tsconfig.json
@@ -12,5 +12,6 @@
   },
   "include": [
     "src/**/*",
+    "dist/**/*.d.ts"
   ]
 }
